move o.a.l.search to o.a.l.sandbox.search; relax classes/methods visibility in the core module
