+++ lucene/core/src/java/org/apache/lucene/search/SloppyPhraseMatcher.java
@@ -53,7 +53,7 @@ import org.apache.lucene.util.FixedBitSet;
  * would get same score as "g f"~2, although "c b"~2 could be matched twice.
  * We may want to fix this in the future (currently not, for performance reasons).
  */
-final class SloppyPhraseMatcher extends PhraseMatcher {
+public final class SloppyPhraseMatcher extends PhraseMatcher {
 
   private final PhrasePositions[] phrasePositions;
 
@@ -81,7 +81,7 @@ final class SloppyPhraseMatcher extends PhraseMatcher {
   private boolean positioned;
   private int matchLength;
 
-  SloppyPhraseMatcher(PhraseQuery.PostingsAndFreq[] postings, int slop, ScoreMode scoreMode, SimScorer scorer, float matchCost, boolean captureLeadMatch) {
+  public SloppyPhraseMatcher(PhraseQuery.PostingsAndFreq[] postings, int slop, ScoreMode scoreMode, SimScorer scorer, float matchCost, boolean captureLeadMatch) {
     super(matchCost);
     this.slop = slop;
     this.numPostings = postings.length;
