+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/document/LongPointMultiRangeBuilder.java
@@ -17,7 +17,7 @@
 
 package org.apache.lucene.sandbox.document;
 
-import org.apache.lucene.search.MultiRangeQuery;
+import org.apache.lucene.sandbox.search.MultiRangeQuery;
 
 import static org.apache.lucene.document.LongPoint.decodeDimension;
 import static org.apache.lucene.document.LongPoint.pack;
