+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/search/LongHashSet.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.search;
+package org.apache.lucene.sandbox.search;
 
 import java.util.AbstractSet;
 import java.util.Arrays;
